Hackathon/Hackathon/requirements.txt.txt
Flask==2.3.2
Pillow==9.5.0
numpy==1.24.3
Werkzeug==2.3.6